refactor : 에러 반환 객체 경로 수정
